use winapi cause winit is not working
Winit would seemingly conflict with whatever wallpaper engine is doing to set the window size, position, etc, and cause wp engine to keep retrying when setting the wallpaper. This would cause the window to flash like 10 times before actually setting the wallpaper.

I decide to just remove winit and make the most basic window with WinAPI as possible to remove any posibilities of conflict, and it seems to have helped.